The selenium files were just removed from main, and this commit preserves them in QA. This also pulls in a few valid updates.
Conflicts:
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/ApiKeysIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/ApplicationDetailPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/BasePage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/DashboardPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/DefectTrackerIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/FindingEditPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/LoginPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/RemoteProvidersIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/RolesIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/ScanDetailPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/ScanIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/TeamDetailPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/TeamIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/UserIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/UserPermissionsPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/VulnerabilityDetailPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/pages/WafIndexPage.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/APIKeysTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/ApplicationTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/BaseTest.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/DashboardTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/DefectTrackerTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/DocumentTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/LoginTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/ReportTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/RoleTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/ScanContents.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/ScanQueueTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/TeamTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/UserTests.java
	threadfix-main/src/test/java/com/denimgroup/threadfix/selenium/tests/WafTests.java
	threadfix-main/src/test/resources/Drivers/IEDriverServer.exe
	threadfix-main/src/test/resources/Drivers/chromedriver
	threadfix-main/src/test/resources/Drivers/chromedriver.exe

+++ threadfix-importers/src/main/java/com/denimgroup/threadfix/importer/impl/AbstractChannelImporter.java
@@ -28,6 +28,7 @@ import com.denimgroup.threadfix.data.dao.ChannelTypeDao;
 import com.denimgroup.threadfix.data.dao.ChannelVulnerabilityDao;
 import com.denimgroup.threadfix.data.dao.GenericVulnerabilityDao;
 import com.denimgroup.threadfix.data.entities.*;
+import com.denimgroup.threadfix.importer.exception.ScanFileUnavailableException;
 import com.denimgroup.threadfix.importer.interop.ChannelImporter;
 import com.denimgroup.threadfix.importer.interop.ScanCheckResultBean;
 import com.denimgroup.threadfix.importer.interop.ScanImportStatus;
@@ -176,14 +177,16 @@ public abstract class AbstractChannelImporter extends SpringBeanAutowiringSuppor
 	public void deleteScanFile() {
 		
 		closeInputStream(inputStream);
-		
-		File file = new File(inputFileName);
-		if (file.exists()) {
-			if (!file.delete()) {
-				log.warn("Scan file deletion failed, calling deleteOnExit()");
-				file.deleteOnExit();
-			}
-		}
+
+        if (shouldDeleteAfterParsing) {
+            File file = new File(inputFileName);
+            if (file.exists()) {
+                if (!file.delete()) {
+                    log.warn("Scan file deletion failed, calling deleteOnExit()");
+                    file.deleteOnExit();
+                }
+            }
+        }
 	}
 
 	protected void deleteZipFile() {
@@ -464,7 +467,7 @@ public abstract class AbstractChannelImporter extends SpringBeanAutowiringSuppor
 	 */
 	protected InputStream getFileFromZip(String fileName) {
 		if (zipFile == null || fileName == null || fileName.trim().equals("")) {
-			return null;
+            throw new ScanFileUnavailableException("zipFile was null or fileName died");
 		}
 		
 		InputStream inputStream = null;
@@ -483,19 +486,20 @@ public abstract class AbstractChannelImporter extends SpringBeanAutowiringSuppor
 	
 	protected ZipFile unpackZipStream() {
 		if (this.inputStream == null) {
-			return null;
+            throw new ScanFileUnavailableException("inputStream was null.");
 		}
 
 		log.debug("Attempting to unpack the zip stream.");
-	
-		diskZipFile = new File("temp-zip-file");
+
+        long timeStamp = new Date().getTime();
+
+		diskZipFile = new File("temp-zip-file" + timeStamp);
 
         if (diskZipFile.exists()) {
             if (!diskZipFile.delete()) {
                 System.out.println("Unable to proceed; can't write to " + diskZipFile.getAbsolutePath());
             }
         }
-
 
 		ZipFile zipFile = null;
 		FileOutputStream out = null;
